refactor: repository폴더 명 변경에 따른 수정
