+++ src/test/java/lotto/service/LottoDrawServiceTest.java
@@ -4,8 +4,7 @@ import lotto.domain.Bonus;
 import lotto.domain.Lotto;
 import lotto.domain.LottoRank;
 import lotto.domain.WinningChecker;
-import lotto.reposi.InMemoryLottoRepository;
-import lotto.service.LottoDrawService;
+import lotto.repository.InMemoryLottoRepository;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
