+++ src/test/java/lotto/service/LottoTicketServiceTest.java
@@ -1,8 +1,8 @@
 package lotto.service;
 
 import lotto.constant.LottoValueConstant;
-import lotto.reposi.InMemoryLottoRepository;
-import lotto.reposi.LottoRepository;
+import lotto.repository.InMemoryLottoRepository;
+import lotto.repository.LottoRepository;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
